+++ docs/changelog.md
@@ -508,6 +508,7 @@ Many feature gates are removed in order to make most features make always availa
 - new types: `Event`, `EventKind`, `EventWindow`.
 - change `EventPointer.pressure` field to be `f32bits_niche`.
 - derive `Eq` & `Hash` for most event types.
+- rename `time_stamp` fields to `timestamp`.
 - update `EventTimestamp`
   - implement `ConstInit`.
   - remove all inner unsafe.